Remote job run. Fix for remote run on REST minions not serializing events. Added Strings to Constants file.

+++ framework/src/main/java/org/mvss/karta/framework/runtime/models/ExecutionStepPointer.java
@@ -25,7 +25,7 @@ public class ExecutionStepPointer implements Serializable
    private String            scenario;
    private String            step;
 
-   private int               iterationIndex;
+   private long              iterationIndex;
    private int               stepIndex;
 
    // private SourcePointer sourcePointer;
